requirements.txt
# --- LLM / LangChain ---
langchain
langchain-openai
langchain-google-genai
langchain-community
openai
google-generativeai

# --- Core Validation ---
pydantic

# --- Environment Variables ---
python-dotenv

# --- UI ---
streamlit

# --- PDF Export ---
reportlab

# --- Database ---
mysql-connector-python